feat: aux functions for lists

+++ src/controllers/lists/index.js
@@ -6,6 +6,7 @@ export const NewListController = (serviceContainer) => {
   const notAllowed = { errors: "user.not-allowed" };
   const boardNotFound = { errors: "board.not-found" };
   const listNotFound = { errors: "list.not-found" };
+  const listDoesntBelongToBoard = { errors: "list.doesnt-belong-to-board" };
 
   const boardExists = async (boardId) => {
     return boardService.exists(boardId);
@@ -13,6 +14,12 @@ export const NewListController = (serviceContainer) => {
 
   const listExists = async (listId) => {
     return listService.exists(listId);
+  };
+
+  const validateList = async (listId, boardId) => {
+    const actualBoardId = await listService.getBoard(listId);
+    if (actualBoardId == boardId) return;
+    return listDoesntBelongToBoard;
   };
 
   const hasPermission = async (boardId, userId) => {
@@ -28,6 +35,7 @@ export const NewListController = (serviceContainer) => {
   const update = async (boardId, userId, listId, name) => {
     if (!(await boardExists(boardId))) return boardNotFound;
     if (!(await listExists(listId))) return listNotFound;
+    if (!(await validateList(listId, boardId))) return listDoesntBelongToBoard;
     if (!(await hasPermission(boardId, userId))) return notAllowed;
     return await listService.update(listId, name);
   };
@@ -35,6 +43,7 @@ export const NewListController = (serviceContainer) => {
   const deleteList = async (boardId, userId, listId) => {
     if (!(await boardExists(boardId))) return boardNotFound;
     if (!(await listExists(listId))) return listNotFound;
+    if (!(await validateList(listId, boardId))) return listDoesntBelongToBoard;
     if (!(await hasPermission(boardId, userId))) return notAllowed;
     return await listService.deleteList(listId);
   };
